fix : 최저 가격의 상품 수량이 0이 되어도 끝나지 않는 오류 수정
