+++ docs/README.md
@@ -56,5 +56,6 @@
 - [x]  product -> model보다 service에서 사용자가 원하는 결과물을 도출하도록 수정
 - [ ]  메소드명, 변수명 더 명확하게 변경
 - [x]  예외 검증 코드 중복 제거
-- [ ]  coins,products 리스트로 변경 => 최저 가격의 상품 수량이 0이 되어도 끝나지 않음
+- [x]  coins 리스트로 변경
+- [x]  최저 가격의 상품 수량이 0이 되어도 끝나지 않는 오류 수정
