+++ src/main/java/vendingmachine/model/Products.java
@@ -1,9 +1,12 @@
 package vendingmachine.model;
 
 import static vendingmachine.constant.Constant.*;
+import static vendingmachine.constant.ErrorMessage.*;
 
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 public class Products {
 
@@ -19,7 +22,25 @@ public class Products {
 		}
 	}
 
-	public List<Product> getProducts() {
-		return products;
+	public int getMinimumPrice() {
+		Comparator<Product> comparatorByPrice = Comparator.comparingInt(Product::getPrice);
+		return products.stream()
+			.min(comparatorByPrice)
+			.orElseThrow(NoSuchElementException::new).getPrice();
+	}
+
+	public boolean isOverZeroAllProductCount() {
+		return 0 < products.stream().mapToInt(Product::getQuantity).sum();
+	}
+
+	public Product findByName(String productName) {
+		return products.stream()
+			.filter(product -> product.getName().equals(productName))
+			.findFirst()
+			.orElseThrow(() -> new IllegalArgumentException(PRODUCT_NOT_EXIST_MSG));
+	}
+
+	public void removeProduct(Product product) {
+		products.remove(product);
 	}
 }
